♻️ refactor: 타임라인 컴포넌트 데스크탑, 모바일 통합
