+++ apps/web/src/components/TimeLine/ThirtyMinutesTimeBox.tsx
@@ -1,8 +1,8 @@
 import useIsMobile from "@src/hooks/useIsMobile";
 import clsx from "clsx";
 
-import { ThirtyMinutesTimeBoxProps } from "../TimeLinetypes";
-import { TimeText } from "../TimeText";
+import { ThirtyMinutesTimeBoxProps } from "./TimeLinetypes";
+import { TimeText } from "./TimeText";
 
 type TimeVerticalLineProps = Pick<
   ThirtyMinutesTimeBoxProps,
@@ -19,8 +19,8 @@ export function TimeVerticalLine({
     <div
       className={clsx(
         "z-10 w-1",
-        isHalfHour ? "h-12" : "h-55",
-        isTestCurrentTime ? "w-1 bg-black" : "bg-gray-200",
+        isHalfHour ? "h-12 md:h-16" : "h-55 md:h-75",
+        isTestCurrentTime ? "bg-black" : "bg-gray-200",
       )}
     />
   );
@@ -49,7 +49,7 @@ function ThirtyMinutesTimeBox({
   const is24Hour = time === "24:00";
 
   return (
-    <li className="flex w-48 flex-col justify-end gap-40">
+    <li className="md:w-75 flex w-48 flex-col justify-end gap-40">
       {/* 상단 시간 텍스트 */}
       {isMobile && (
         <TimeText
@@ -60,7 +60,7 @@ function ThirtyMinutesTimeBox({
         />
       )}
       {/* 세로 구분선과 점선을 포함하는 30분 단위 박스 */}
-      <div className="h-55 relative flex w-48 flex-col justify-end">
+      <div className="h-55 md:h-75 relative flex w-48 flex-col justify-end md:w-72">
         {/* 24:00이 아닐 때만 호버 효과 적용 */}
         {!is24Hour && (
           <div className="absolute inset-0 left-1 cursor-pointer hover:bg-[#ede2f9] hover:shadow-[inset_0_0_0_1px_#e0c8fa]">
